separated newsAlerts and alerts to make further development less messy
