+++ server/backend/news.go
@@ -2,7 +2,6 @@ package backend
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	pb "github.com/TUM-Dev/Campus-Backend/server/api/tumdev"
 	"github.com/TUM-Dev/Campus-Backend/server/model"
@@ -10,8 +9,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/timestamppb"
-	"gorm.io/gorm"
 )
 
 func (s *CampusServer) GetNewsSources(ctx context.Context, _ *emptypb.Empty) (newsSources *pb.NewsSourceReply, err error) {
@@ -35,27 +32,4 @@ func (s *CampusServer) GetNewsSources(ctx context.Context, _ *emptypb.Empty) (ne
 		})
 	}
 	return &pb.NewsSourceReply{Sources: resp}, nil
-}
-
-func (s *CampusServer) GetTopNews(ctx context.Context, _ *emptypb.Empty) (*pb.GetTopNewsReply, error) {
-	if err := s.checkDevice(ctx); err != nil {
-		return nil, err
-	}
-
-	var res *model.NewsAlert
-	err := s.db.Joins("Files").Where("NOW() between `from` and `to`").First(&res).Error
-	if errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, status.Error(codes.NotFound, "no currenty active top news")
-	} else if err != nil {
-		log.WithError(err).Error("could not GetTopNews")
-		return nil, status.Error(codes.Internal, "could not GetTopNews")
-	}
-
-	return &pb.GetTopNewsReply{
-		ImageUrl: res.Files.URL.String,
-		Link:     res.Link.String,
-		Created:  timestamppb.New(res.Created),
-		From:     timestamppb.New(res.From),
-		To:       timestamppb.New(res.To),
-	}, nil
 }
